+++ src/Monitor/Monitor/ActionGroups/SetAzureRmActionGroupCommand.cs
@@ -114,6 +114,10 @@ namespace Microsoft.Azure.Commands.Insights.ActionGroups
                 if (ParameterSetName == ByInputObject)
                 {
                     this.ResourceGroupName = this.InputObject.ResourceGroupName;
+                    if (this.Name == null)
+                    {
+                        this.Name = this.InputObject.Name;
+                    }
                     if (this.ShortName == null)
                     {
                         this.ShortName = this.InputObject.GroupShortName;