+++ src/Sql/Sql/ChangeLog.md
@@ -18,11 +18,15 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed a bug in Data Classification cmdlets.
+
+## Version 2.7.0
 * Added UsePrivateLinkConnection to `New-AzSqlSyncGroup`, `Update-AzSqlSyncGroup`, `New-AzSqlSyncMember` and `Update-AzSqlSyncMember`
 * Added SyncMemberAzureDatabaseResourceId to `New-AzSqlSyncMember` and `Update-AzSqlSyncMember`
 * Added Guest user lookup support to Set SQL Server Azure Active Directory Admin cmdlet
 * Remove IsAzureADOnlyAuthentication parameter from Set-AzSqlServerActiveDirectoryAdministrator as it is not usable. 
 * Added support for service principal for Set SQL Server Azure Active Directory Admin cmdlet
+
 
 ## Version 2.6.1
 * Enhance performance of: